Add the ability to "clear" the number of restarts for a process.
When a proc hits the max number of restarts, the system will mark it as having done so and not restart it. This new capability allows a user to "reset" the number of restarts on a proc so the system will restart it, thus allowing it to run until it once again hits max restarts.



git-svn-id: http://svn.open-mpi.org/svn/orcm/trunk@164 e75c48be-c379-0410-a98d-d1e110c868e2